coreclr: add missing runtime properties and other misc. fixes
- the coreclr addon now sets the following properties,
which are used for reflection and dynamic assembly loading:
* APP_CONTEXT_BASE_DIRECTORY => dirname(asmPath)
* APP_PATHS => dirname(asmPath)
* APP_CONTEXT_DEPS_FILES => asmPath.replace_extension() + ".deps.json" (prepended to the list)

this fixes #2

- fixed conversion of cygwin paths to windows paths:
converting /cygdrive/c/foo.exe is the same as /cygdrive/c/foo, and both yield C:\foo.exe (probably to mimic unix).
to workaround this, the code now converts the dirname, and appends the stem to the result

- normalize '/' to '\' (for msys paths)
they both work equally, but '\' is the canonical Windows path separator and should be preferred.

- refactoring to use std::optional instead of pointers in dotnet_init_params